pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "google-genai>=1.46.0",
    "openai>=2.6.0",
    "streamlit-ace>=0.1.1",
    "streamlit>=1.50.0",
    "jedi>=0.19.2",
    "pygments>=2.19.2",
    "psutil>=7.1.1",
    "pydantic>=2.12.3",
]
